Fixed #27829 -- Deprecated settings.DEFAULT_CONTENT_TYPE.
